refactor : String Name 원시값 포장
